New Development: No             JIRA Issue: No       Ready to Release: Yes      Interface Changes: No What Interface Changed:          Test Programs: tArrayOpsDiffShapes   Put in Release Notes: No              Module(s): casa            Description: Rechecking in a compiler warning fix.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/casacore/alma-evla@8525 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ casa/casa/Arrays/ArrayOpsDiffShapes.tcc
@@ -42,7 +42,7 @@ LogicalArray reformedMask(const Array<T>& data, const T truthvalue,
      && data.shape() == desiredform){
     return (data == truthvalue);
   }
-  else if(data.nelements() == desiredform.product()){
+  else if(static_cast<Int>(data.nelements()) == desiredform.product()){
     return (data == truthvalue).reform(desiredform);
   }
   else{
